+++ packages/aesthetic/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.4.1 - 05/09/18
+#### 🐞 Fixed
+* Fixed a bug in which `themeName` was being passed from the HOC.
+
 # 2.4.0 - 05/08/18
 #### 🚀 New
 * Added new `passThemeProp` and `passThemeNameProp` options to control which props are passed to the underlying component.